rag_chatbot/requirements.txt
mcp
qdrant-client
sentence-transformers
torch
numpy
markdown
pyyaml

streamlit
watchdog
